% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_utils.R
\name{arg_in_choices}
\alias{arg_in_choices}
\title{Argument Verification}
\usage{
arg_in_choices(arg, choices, error_message = NULL, call_ = F)
}
\arguments{
\item{arg}{Character: passed-in argument name}

\item{choices}{Character vector: a character vector of candidate values}

\item{error_message}{Character: user-defined error message}

\item{call_}{Logical: logical, indicating if the call should become part of
the error message. See \code{call.} in \code{\link[base]{stop}}}
}
\value{
Error message if test fails
}
\description{
Tests if a passed-in argument takes a value from a table of
 candidate values and returns an error message if not.
}
\examples{
alignment <- 'left'
arg_in_choices(alignment, c('left', 'center', 'right'))
}
