% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemap.R
\name{basemap}
\alias{basemap}
\title{basemap}
\usage{
basemap(countryname = "world")
}
\arguments{
\item{countryname}{string, name of base map area.}
}
\value{
sf object
}
\description{
Function to return base map data
}
