% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_utils.R
\name{check_var_class}
\alias{check_var_class}
\title{Variable Class Test}
\usage{
check_var_class(var, class_method, class_name, error_message = NULL)
}
\arguments{
\item{var}{Object: an object for class test}

\item{class_method}{Function vector: class test function, e.g.,
\code{\link[base]{is.numeric}}, \code{\link[base]{is.logical}} and etc.}

\item{class_name}{Character: name of the class}

\item{error_message}{Character: user-defined error message}
}
\value{
Error message if test fails
}
\description{
Tests if a variable of a certain class and returns an error
 message if not
}
\examples{
obj <- 1:10
check_var_class(obj, is.numeric, 'numeric')
}
