% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_intervals.R
\name{classify_intervals}
\alias{classify_intervals}
\title{classify_intervals}
\usage{
classify_intervals(
  data,
  value,
  breaks,
  sym_start = "(",
  sym_sep = " to ",
  sym_end = "]"
)
}
\arguments{
\item{data}{data.frame, input data.frame}

\item{value}{string, variable name of value to classify}

\item{breaks}{vector numeric, class breaks}

\item{sym_start}{string, start of interval label. Default: "(".}

\item{sym_sep}{string, separator of interval label. Default: "to ".}

\item{sym_end}{string, end of interval label. Default: "]".}
}
\value{
data.frame
}
\description{
Function to classify data in custom intervals
}
