% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_utils.R
\name{column_in_dataframe}
\alias{column_in_dataframe}
\title{Test If A Column Is Found in A Data Frame}
\usage{
column_in_dataframe(df, column, error_message = NULL, call_ = F)
}
\arguments{
\item{df}{List/Data frame}

\item{column}{Character: A column name}

\item{error_message}{Character: user-defined error message}

\item{call_}{Logical: logical, indicating if the call should become part of
the error message. See \code{call.} in \code{\link[base]{stop}}}
}
\value{
Error message if test fails
}
\description{
A function that check if \code{column} is one of \code{df}'s
 columns and returns an error message if not.
}
\examples{
test_df <- data.frame(a = 1:10, b = 11:20)
column_in_dataframe(test_df, 'a')
\dontrun{
column_in_dataframe(test_df, 'c')
}
}
