% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lims.R
\name{geo_lims}
\alias{geo_lims}
\title{geo_lims}
\usage{
geo_lims(object)
}
\arguments{
\item{object}{data.frame, Combined mobility data.}
}
\value{
list
}
\description{
Define plot lims using a spatial object
}
