% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_add.R
\name{gg_add_text}
\alias{gg_add_text}
\title{Add Text to a Plot}
\usage{
gg_add_text(ggplot_obj, x, y = NULL, text, on_border = FALSE,
  panel_recycling = FALSE, ...)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}

\item{x, y}{the x and y co-ordinates to be used to position the legend. They
can be specified by keyword or in any way which is accepted by
\code{\link[grDevices]{xy.coords}}. See the details of
\code{\link[graphics]{legend}}}

\item{text}{a character vector or \code{\link[base]{expression}} specifying
the text to be written.}

\item{on_border}{when the location is specified by setting \code{x} to a
single keyword from the list "bottomright", "bottom", "bottomleft", "left",
"topleft", "top", "topright", "right" and "center", \code{text} can be
placed right adjacent to the border (`on_border = TRUE`) or placed with
proper space away from the border (`on_border = FALSE`)}

\item{panel_recycling}{When \code{text} is added to a plot with panel
structure, i.e. either \code{\link[ggplot2]{facet_grid}} or
\code{\link[ggplot2]{facet_grid}} is used in building the ggplot2 object,
and lengths of \code{x}, \code{y}, and \code{text} are smaller than the
number of panels, \code{text} and its corresponding location will be
recycled until the last panel if `panel_recycling = TRUE`}

\item{...}{more arguments passed to \code{\link[ggplot2]{geom_text}}}
}
\description{
Add text to a \href{http://ggplot2.org/}{ggplot2} object in a similar way to
 \code{\link[graphics]{legend}}
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
p_with_text <- gg_add_text(
    p, 'topright',
    text = paste('Correlation:', round(cor(mtcars$wt, mtcars$mpg), 3))
)

library(dplyr)
q <- ggplot(data = diamonds, aes(depth, price)) + geom_point() +
    facet_grid(~ cut)
corr_df <- diamonds \%>\%
    group_by(cut) \%>\%
    summarise(corr = paste('Correlation:', round(cor(price, depth), 3)))
q_with_text <- gg_add_text(
    q, 'top', text = corr_df$corr, on_border = TRUE
)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
