% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_has.R
\name{gg_has_facet_grid}
\alias{gg_has_facet_grid}
\title{Whether Has Facetting Grid}
\usage{
gg_has_facet_grid(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
\code{TRUE} if \code{\link[ggplot2]{facet_grid}} is used;
 \code{FALSE} otherwise
}
\description{
Check if a ggplot2 object has facetting grid structure, i.e.
 if \code{\link[ggplot2]{facet_grid}} is used in building the ggplot2 object
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) + geom_point()
print(gg_has_facet_grid(p))
p <- p + facet_grid(. ~ cyl)
print(gg_has_facet_grid(p))
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
