% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_has.R
\name{gg_has_facet_layout}
\alias{gg_has_facet_layout}
\title{Whether Has Facetting Panels}
\usage{
gg_has_facet_layout(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
\code{TRUE} if either \code{\link[ggplot2]{facet_grid}} or
 \code{\link[ggplot2]{facet_grid}}is used; \code{FALSE} otherwise
}
\description{
Check if a ggplot2 object has facetting panel structure, i.e.
 if either \code{\link[ggplot2]{facet_grid}} or
 \code{\link[ggplot2]{facet_grid}} is used in building the ggplot2 object
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) + geom_point()
print(gg_has_facet_grid(p))
print(gg_has_facet_layout(p + facet_grid(. ~ cyl)))
print(gg_has_facet_layout(p + facet_wrap(~ cyl)))
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
