% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_has.R
\name{gg_has_facet_wrap}
\alias{gg_has_facet_wrap}
\title{Whether Has Facetting Ribbon}
\usage{
gg_has_facet_wrap(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
\code{TRUE} if \code{\link[ggplot2]{facet_wrap}} is used;
 \code{FALSE} otherwise
}
\description{
Check if a ggplot2 object has facetting ribbon structure, i.e.
 if \code{\link[ggplot2]{facet_wrap}} is used in building the ggplot2 object
}
\examples{
p <- ggplot(mpg, aes(displ, hwy)) + geom_point()
print(gg_has_facet_wrap(p))
p <- p + facet_wrap(~ cyl + drv)
print(gg_has_facet_wrap(p))
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
