% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_remove.R
\name{gg_remove}
\alias{gg_remove}
\title{Remove Plot Elements}
\usage{
gg_remove(ggplot_obj, elements_to_remove = NULL)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}

\item{elements_to_remove}{String vector: a character vector dictating which
elements(s) to be removed from \code{ggplot_obj} with elements from:
'xlab', 'ylab', 'title', 'subtitle', 'legend_title', 'xaxis', 'yaxis',
'legend', 'grid', 'grid_x', 'grid_y', 'xticks', 'yticks'. By default,
nothing gets removed, i.e. same \code{ggplot_obj} gets returned.}
}
\value{
A ggplot2 object
}
\description{
Remove one or more elements from a \href{http://ggplot2.org/}{ggplot2} Object
}
\examples{
p <- ggplot(mtcars, aes(mpg, wt, colour = cyl)) + geom_point() +
    labs(colour = 'Cylinders',
         x = 'Miles/(US) gallon',
         y = 'Weight (1000 lbs)',
         title = 'New plot title',
         subtitle = 'A subtitle',
         caption = '(based on data from ...)')
new_p <- gg_remove(p, c('title', 'caption'))
p_no_xlab <- gg_remove_xlab(p)
p_no_legend_title <- gg_remove_legend_title(p)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
