% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_facet_layout}
\alias{gg_retrieve_facet_layout}
\title{Retrieve Facetting Layout}
\usage{
gg_retrieve_facet_layout(ggplot_obj, facet_var_only = TRUE)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
