% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_facet_row}
\alias{gg_retrieve_facet_row}
\title{Retrieve Row Facetting Variable}
\usage{
gg_retrieve_facet_row(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
Row facetting variable. `NULL` is returned if
 \code{\link[ggplot2]{facet_grid}} is not used in building the ggplot2
 object
}
\description{
Get the name of the column that is mapped to row facetting
 variable of a ggplot2 object through \code{\link[ggplot2]{facet_grid}}
}
\examples{
p <- ggplot(mpg, aes(displ, cty)) + geom_point() + facet_grid(. ~ cyl)
facet_row <- gg_retrieve_facet_row(p)
print(facet_row)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
