% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_facet_wrap}
\alias{gg_retrieve_facet_wrap}
\title{Retrieve Wrap Panel Variable(s)}
\usage{
gg_retrieve_facet_wrap(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
Wrap panel variable(s). `NULL` is returned if
 \code{\link[ggplot2]{facet_wrap}} is not used in building the ggplot2
 object
}
\description{
Get the name(s) of the variable(s) passed to the `facets`
 argument of \code{\link[ggplot2]{facet_wrap}} that wraps a 1d sequence
 of panels into 2d. Refer to \code{\link[ggplot2]{facet_wrap}} for the
 details
}
\examples{
p <- ggplot(mpg, aes(displ, hwy)) + geom_point() + facet_wrap(~ cyl + drv)
facet_wrap <- gg_retrieve_facet_wrap(p)
print(facet_wrap)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
