% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_range}
\alias{gg_retrieve_range}
\title{Retrieve Axis Ranges from A ggplot2 Object}
\usage{
gg_retrieve_range(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
A dataframe, which always contains the following columns
 * `x_min`: lower limit of x-axis
 * `x_max`: upper limit of x-axis
 * `y_min`: lower limit of y-axis
 * `y_max`: upper limit of y-axis
 In the case of facetting panels, i.e. with the use of either
 \code{\link[ggplot2]{facet_grid}} or \code{\link[ggplot2]{facet_wrap}},
 besides the above four, additional columns of facetting variable(s) are
 included as well.
}
\description{
Get the actual x- and y-axis plotting ranges from a ggplot2
 object
}
\examples{
# axis range of a scatter plot
scatter_plot <- ggplot(mtcars, aes(wt, mpg)) + geom_point()
print(gg_retrieve_range(scatter_plot))

# axis range of an errorbar plot
df <- data.frame(
    trt = factor(c(1, 1, 2, 2)),
    resp = c(1, 5, 3, 4),
    group = factor(c(1, 2, 1, 2)),
    upper = c(1.1, 5.3, 3.3, 4.2),
    lower = c(0.8, 4.6, 2.4, 3.6)
)
errorbar_plot <- ggplot(df, aes(trt, resp, colour = group)) +
    geom_errorbar(aes(ymin = lower, ymax = upper), width = 0.2)
print(gg_retrieve_range(errorbar_plot))

# axis range of a bar chart
bar_chart <- ggplot(mpg, aes(class)) + geom_bar()
print(gg_retrieve_range(bar_chart))

# axis range of panels in a grid
facet_col_plot <- ggplot(mpg, aes(displ, cty)) + geom_point() +
    facet_grid(. ~ cyl)
print(gg_retrieve_range(facet_col_plot))

# axis range of panels in a 2d grid
facet_2d_plot <- ggplot(mpg, aes(displ, cty)) + geom_point() +
    facet_grid(drv ~ cyl)
print(gg_retrieve_range(facet_2d_plot))
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
