% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_title}
\alias{gg_retrieve_title}
\title{Retrieve Plot Title}
\usage{
gg_retrieve_title(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
Plot title
}
\description{
Get plot title of a ggplot2 object
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point(aes(colour = factor(cyl)))
plot_title <- gg_retrieve_title(p)
print(plot_title)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
