% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_x}
\alias{gg_retrieve_x}
\title{Retrieve X-axis Variable}
\usage{
gg_retrieve_x(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
A column name
}
\description{
Get the name of the column that is mapped to the x-axis
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point(aes(colour = factor(cyl)))
x <- gg_retrieve_x(p)
print(x)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
