% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_x_lab}
\alias{gg_retrieve_x_lab}
\title{Retrieve X-axis Label}
\usage{
gg_retrieve_x_lab(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
X-axis label
}
\description{
Get x-axis label of a ggplot2 object
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point(aes(colour = factor(cyl)))
x_label <- gg_retrieve_x_lab(p)
print(x_label)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
