% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_y}
\alias{gg_retrieve_y}
\title{Retrieve Y-axis Variable}
\usage{
gg_retrieve_y(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
A column name
}
\description{
Get the name of the column that is mapped to the y-axis
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point(aes(colour = factor(cyl)))
y <- gg_retrieve_y(p)
print(y)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
