% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_retrieve.R
\name{gg_retrieve_y_lab}
\alias{gg_retrieve_y_lab}
\title{Retrieve Y-axis Label}
\usage{
gg_retrieve_y_lab(ggplot_obj)
}
\arguments{
\item{ggplot_obj}{a ggplot2 object}
}
\value{
Y-axis label
}
\description{
Get y-axis label of a ggplot2 object
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg)) + geom_point(aes(colour = factor(cyl)))
y_label <- gg_retrieve_y_lab(p)
print(y_label)
}
\author{
Feiyang Niu (Feiyang.Niu@gilead.com)
}
