% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggutils.R
\name{ggbrush}
\alias{ggbrush}
\title{Brush ggplot2}
\usage{
ggbrush(plotExpr, direction = c("xy", "x", "y"))
}
\description{
Call ggbrush with a ggplot2 object, and the dimensions which
should be brushed (try "xy" for scatter, "x" for histogram).
The plot will show in RStudio Viewer or your web browser, and
any observations selected by the user will be returned.
}

