% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggutils.R
\name{ggidentify}
\alias{ggidentify}
\title{Identify points on a two-dimensional ggplot2}
\usage{
ggidentify(plotExpr, threshold = 5, maxpoints = 1, ...)
}
\description{
Identify points on a two-dimensional ggplot2
}

