% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsave_png_pdf.R
\name{ggsave_png_pdf}
\alias{ggsave_png_pdf}
\title{Save a pdf and png image with identical path and filename}
\usage{
ggsave_png_pdf(p, path, width = 8.5, height = 6)
}
\arguments{
\item{p}{ggplot object, ggplot object to be saved}

\item{path}{character, output pdf filepath}

\item{width}{numeric, output figure width (inches)}

\item{height}{numeric, output figure height (inches)}
}
\value{
none
}
\description{
Save a pdf and png image with identical path and filename
}
