% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggutils.R
\name{ggzoom}
\alias{ggzoom}
\title{Zoom in on ggplots}
\usage{
ggzoom(plotExpr)
}
\arguments{
\item{plotExpr}{A ggplot2 plot object}
}
\value{
The final zoom bounds (or \code{NULL} if unzoomed)
}
\description{
Zoom in on a ggplot object by dragging a selection with the mouse.
}
\examples{
\donttest{
library(ggplot2)
p <- ggplot(cars, aes(speed, dist)) + geom_point()
ggzoom(p)

# Also works for one-dimensional visualizations
p2 <- ggplot(cars, aes(dist)) + geom_histogram()
gzoom(p2)
}
}

