% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusive_class_breaks.R
\name{inclusive_class_breaks}
\alias{inclusive_class_breaks}
\title{inclusive_class_breaks}
\usage{
inclusive_class_breaks(var, style, n, offset = 1e-07)
}
\arguments{
\item{var}{numeric, input variable}

\item{style}{string, "style" of class breaks}

\item{n}{numeric, number of classes}

\item{offset}{numeric, offset added/subtracted from either end of \code{var}. Makes class breaks inclusive.}
}
\value{
numeric
}
\description{
Define class breaks using "styles" supported by \code{classInt::classIntervals()}. Inclusive breaks include the highest and lowest values in a variable.
}
