% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_utils.R
\name{is_blank}
\alias{is_blank}
\title{Test If An Object Is Empty}
\usage{
is_blank(obj, empty_str_triggers = TRUE, false_triggers = FALSE)
}
\arguments{
\item{obj}{Object: an object to be tested}

\item{empty_str_triggers}{Logical: \code{TRUE} (default) treats empty string
(i.e. \code{''}) as empty and \code{FALSE} otherwise}

\item{false_triggers}{Logical: \code{TRUE} treats logical value \code{FALSE}
as empty and \code{FALSE} (default) otherwise}
}
\value{
A logical value that indicates whether the tested object is empty or
 not
}
\description{
An empty object is any of the following: \code{NULL}, \code{NA}
 or vector of \code{NA}s, Object of length \code{0}, \code{''} if
 \code{empty_str_triggers} equals to \code{TRUE}, and \code{FALSE} if
 \code{false_triggers} equals to \code{TRUE}
}
\examples{
is_blank(NA)
is_blank('')
}
