% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_intervals.R
\name{label_interval}
\alias{label_interval}
\title{classify_intervals}
\usage{
label_interval(breaks, sym_start, sym_sep, sym_end)
}
\arguments{
\item{breaks}{vector numeric, class breaks}

\item{sym_start}{string, start of interval label. Default: "(".}

\item{sym_sep}{string, separator of interval label. Default: "to ".}

\item{sym_end}{string, end of interval label. Default: "]".}

\item{data}{data.frame, input data.frame}

\item{value}{string, variable name of value to classify}
}
\value{
data.frame
}
\description{
Function to classify data in custom intervals
}
