% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_basemap.R
\name{plot_basemap}
\alias{plot_basemap}
\title{plot_basemap}
\usage{
plot_basemap(
  countryname = NULL,
  world_fill = "transparent",
  world_size = 0.05,
  world_color = "black",
  country_fill = "transparent",
  country_size = 0.1,
  country_color = "black"
)
}
\arguments{
\item{countryname}{string, name of basemap area.}

\item{world_fill}{string, fill color of world data. Defaults to 'transparent'.}

\item{world_size}{float, line thickness of world data. Defaults to 0.5.}

\item{world_color}{string, line color of world data. Defaults to 'black'.}

\item{country_fill}{string, fill color of country data. Defaults to 'transparent'.}

\item{country_size}{float, line thickness of country data. Defaults to 0.5.}

\item{country_color}{string, line color of country data. Defaults to 'black'.}
}
\value{
list
}
\description{
Plot a gloabl basemap with optional country focus map
}
