% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualitative_pal.R
\name{qualitative_pal}
\alias{qualitative_pal}
\title{qualitative_pal}
\usage{
qualitative_pal(labels)
}
\arguments{
\item{labels}{string vector, name to be symbolized.}
}
\value{
named character vector.
}
\description{
Function to define a palette of distinct colors
}
