% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_ci.R
\name{summarise_ci}
\alias{summarise_ci}
\title{summarise_ci}
\usage{
summarise_ci(
  data,
  group_variables,
  summary_variable,
  ci = c(0.1, 0.9, 0.4, 0.6, 0.25, 0.75)
)
}
\arguments{
\item{data}{data.frame, data to be summarised.}

\item{group_variables}{character vector, grouping variables.}

\item{summary_variable}{character, variable to be sumamrised.}

\item{ci}{float vector, ci intervals.}
}
\value{
sf object
}
\description{
Function to summarise a target variable given a vector of density intervals by group
}
