% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_lims.R
\name{geo_lims}
\alias{geo_lims}
\title{geo_lims}
\usage{
geo_lims(object, zoom_x = 0, zoom_y = 0)
}
\arguments{
\item{object}{data.frame, Combined mobility data.}

\item{zoom_x}{x zoom}

\item{zoom_y}{y zoom}
}
\value{
list
}
\description{
Define plot limits using a spatial object
}
