% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggutils.R
\name{ggbrush}
\alias{ggbrush}
\title{Brush ggplot2}
\usage{
ggbrush(plotExpr)
}
\description{
Call ggbrush with a ggplot2 object. The plot will show in RStudio Viewer or
your web browser, and any observations selected by the user will be returned.
}

