% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_interval.R
\name{label_interval}
\alias{label_interval}
\title{label_interval}
\usage{
label_interval(breaks, sym_start, sym_sep, sym_end)
}
\arguments{
\item{breaks}{vector numeric, class breaks}

\item{sym_start}{string, start of interval label.}

\item{sym_sep}{string, separator of interval label.}

\item{sym_end}{string, end of interval label.}
}
\value{
string
}
\description{
Function to label a single interval with custom text. Called internally by classify_intervals.
}
